backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0

# Templates
jinja2==3.1.2

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1